pull gas limit computation to common crate
